refactor: only use Snappy on clipboard contents
The majority of messages sent between client and server are too small to be compressed, and the compression was actually making the messages larger (due to frames and headers).

The client and server can now choose whether to compress the clipboard when sending it, and they must let the recipient know via a flag in the update message. This should not affect the final contents of the clipboard on the system.